Ajustado projeto para ter camadas de Service e Repository.
