+++ src/components/layout/Header.tsx
@@ -1,7 +1,4 @@
-import React from "react";
-import { MapPin, Map } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import { useNavigate } from "react-router-dom";
+import { MapPin } from "lucide-react";
 
 interface HeaderProps {
   title?: string;
@@ -14,16 +11,10 @@ const Header = ({
   logoSize = 24,
   onLogoClick = () => console.log("Logo clicked"),
 }: HeaderProps) => {
-  const navigate = useNavigate();
-  
   const handleLogoClick = () => {
     onLogoClick();
   };
-  
-  const handleMapViewClick = () => {
-    navigate("/map");
-  };
-  
+
   return (
     <header className="w-full h-16 px-4 flex items-center justify-between shadow-sm bg-white border-b border-gray-200">
       <div
@@ -36,19 +27,6 @@ const Header = ({
       >
         <MapPin className="text-blue-600" size={logoSize} />
         <h1 className="text-xl font-bold text-blue-800">{title}</h1>
-      </div>
-
-      <div className="flex items-center gap-2">
-        <Button
-          variant="outline"
-          size="sm"
-          className="text-blue-700 border-blue-700 hover:bg-blue-50 flex items-center gap-2"
-          onClick={handleMapViewClick}
-          aria-label="Visualizar mapa de eventos"
-        >
-          <Map className="h-4 w-4" />
-          Visualizar Mapa
-        </Button>
       </div>
     </header>
   );
