+++ src/types/geocoding.ts
@@ -1,0 +1,11 @@
+export interface GeocodingResult {
+  lat: string;
+  lon: string;
+  display_name: string;
+  error?: string;
+}
+
+export interface Coordinates {
+  latitude: number;
+  longitude: number;
+}
